requirements_test.txt
pytest>=8.0
pytest-asyncio>=0.23
pytest-homeassistant-custom-component>=0.13
aiousbwatcher>=1.0
pyserial>=3.5
